@@ -2,16 +2,16 @@ name: app
 description: Flutter Architecture Blueprints project.
 version: 1.0.0+1
 
-publish_to: 'none' # Remove this line if you wish to publish to pub.dev
+publish_to: 'none'
 
 environment:
-  sdk: '>=2.10.0 <3.0.0'
-  flutter: '>=1.21.0'
+  sdk: '>=2.11.0 <3.0.0'
+  flutter: '>=1.25.0'
 
 dependency_overrides:
   # For DateTime and l10n
   # TODO: Workaround https://github.com/FirebaseExtended/flutterfire/issues/4342
-  intl: ^0.17.0-nullsafety.2
+  intl: ^0.17.0
 
 dependencies:
   flutter:
@@ -20,26 +20,29 @@ dependencies:
     sdk: flutter
 
   # For Google and Firebase
-  firebase_core: ^0.5.3
-  firebase_auth: ^0.18.4+1
+  firebase_core: ^0.7.0
+  firebase_auth: ^0.20.0+1
   google_sign_in: ^4.5.9
-  firebase_crashlytics: ^0.2.3
-  firebase_performance: ^0.4.2
+  firebase_crashlytics: ^0.4.0+1
+  firebase_performance: ^0.5.0+1
 
   # For Architecture
   flutter_hooks: ^0.15.0
-  hooks_riverpod: ^0.12.1
+  hooks_riverpod: ^0.12.4
 
   # For Networking
   dio: ^3.0.10
-  dio_http_cache: ^0.2.11
-  dio_firebase_performance: ^0.2.0
+#  dio_http_cache: ^0.2.11
+  dio_firebase_performance: ^0.3.0
   # For User-Agent Client Hints
   ua_client_hints: ^1.0.3
 
   # For Model
   json_serializable: ^3.5.0
   freezed_annotation: ^0.12.0
+
+  # For DateTime
+  intl: ^0.17.0
 
   # For Key-Value local storage
   shared_preferences: ^0.5.12+4
@@ -51,20 +54,19 @@ dependencies:
   gap: ^1.2.0
   cupertino_icons: ^1.0.0
   google_fonts: ^1.1.1
-  cached_network_image: ^2.5.0
-  flutter_svg: ^0.19.2+1
+  flutter_svg: ^0.20.0-nullsafety.3
 
   # For Utilities
-  get: ^3.24.0
+  get: ^3.25.4
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
 
-  build_runner: ^1.10.12
+  build_runner: ^1.11.1
 
   # For Model
-  freezed: ^0.12.6
+  freezed: ^0.12.7
 
   # For Analyzer
   effective_dart: ^1.3.0
