@@ -1,7 +1,6 @@
 import 'dart:math';
 
 import 'package:dio/dio.dart';
-import 'package:dio_http_cache/dio_http_cache.dart';
 import 'package:flutter/foundation.dart';
 
 import '../../constants.dart';
@@ -16,24 +15,20 @@ class NewsDataSourceImpl implements NewsDataSource {
 
   @override
   Future<News> getNews() {
-    return _dio
-        .get<Map<String, dynamic>>(
-          '/v2/everything',
-          queryParameters: <String, String>{
-            'q': ['anim', 'manga'][Random().nextInt(2)], // For checking reload.
-            'from': DateTime.now()
-                .subtract(
-                  const Duration(days: 28),
-                )
-                .toLocal()
-                .formatYYYYMMdd(),
-            'sortBy': 'publishedAt',
-            'language': 'en',
-            'apiKey': Constants.of().apiKey,
-          },
-          // In-memory cache time-to-live
-          options: buildCacheOptions(const Duration(seconds: 5)),
-        )
-        .then((response) => News.fromJson(response.data));
+    return _dio.get<Map<String, dynamic>>(
+      '/v2/everything',
+      queryParameters: <String, String>{
+        'q': ['anim', 'manga'][Random().nextInt(2)], // For checking reload.
+        'from': DateTime.now()
+            .subtract(
+              const Duration(days: 28),
+            )
+            .toLocal()
+            .formatYYYYMMdd(),
+        'sortBy': 'publishedAt',
+        'language': 'en',
+        'apiKey': Constants.of().apiKey,
+      },
+    ).then((response) => News.fromJson(response.data));
   }
 }
