@@ -1,4 +1,3 @@
-import 'package:cached_network_image/cached_network_image.dart';
 import 'package:flutter/widgets.dart';
 
 import '../../gen/assets.gen.dart';
@@ -6,5 +5,5 @@ import '../../gen/assets.gen.dart';
 ImageProvider loadProfileImage(String imageUrl) {
   return imageUrl == null || imageUrl.isEmpty
       ? Assets.images.iconPlaceholder
-      : CachedNetworkImageProvider(imageUrl);
+      : Image.network(imageUrl);
 }
