@@ -1,5 +1,5 @@
 import 'package:flutter/foundation.dart';
-import 'package:hooks_riverpod/all.dart';
+import 'package:hooks_riverpod/hooks_riverpod.dart';
 
 final loadingStateProvider =
     ChangeNotifierProvider((ref) => LoadingStateViewModel());
